requirements.txt
discord.py
aiohttp
python-dotenv
requests
tzdata
google-api-python-client
google-auth
google-auth-oauthlib
Pillow